+++ objc2-encode/CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   compatibility; use these instead of `c_long::ENCODING` and
   `c_ulong::ENCODING`.
 
+### Changed
+* **BREAKING**: Sealed the `EncodeArguments` trait.
+
 
 ## 2.0.0-beta.2 - 2022-01-03